+++ test/integration/src/pages/redux/Hydration/Hydration.js
@@ -5,7 +5,6 @@ import ReactRenderer from 'soya/lib/page/react/ReactRenderer.js';
 import register from 'soya/lib/client/Register';
 import ReduxStore from 'soya/lib/data/redux/ReduxStore.js';
 import UserProfile from '../../../components/contextual/UserProfile/UserProfile.js';
-import { DevTools, DebugPanel, LogMonitor } from 'redux-devtools/lib/react';
 
 // TODO: Figure out how to do promise polyfill.
 import style from '../../../shared/sitewide.css';
@@ -53,9 +52,6 @@ class Component extends React.Component {
       <UserProfile context={this.props.context} username={'jedikiller'}></UserProfile>
       <h3>Profile 2</h3>
       <UserProfile context={this.props.context} username={'willywonka'}></UserProfile>
-      <DebugPanel top right bottom>
-        <DevTools store={this.props.context.store._store} monitor={LogMonitor} />
-      </DebugPanel>
     </div>
   }